Added NukeX to the "Keeper" interface.
utilities/keeper/software.py
Added NukeX functions.
It's really the same as running 'nuke --nukex'.
The example links to the Nuke example.

icons/software/nukex.png
The .PNG for NukeX, this is really the Nuke icon, it's just a copy, I couldn't find a NukeX icon to match the other one.

git-svn-id: https://cgru.svn.sourceforge.net/svnroot/cgru@2245 0f86f178-d61e-0410-9d12-f6255c3815b8