NEW NETWORK PROCOL VERSION!!!
Script that gets revison now tries to get git sha at first, and on fail tries to get svn revision as did it before.
This script used to #DEFINE revision for sources build, and for export script to store it in a text file.
Cliets does not send revision number to server ( - new protocol version).
So now revision just stored in binaries (and printed at each start).