Introduce automatic reconnection of old renders to new server
When the server goes down, and later on up again, the renders keep on running their
task and claim them against the new server.
This require not to mess up with the render info stored in persistent storage
(/var/tmp/afanasy or the likes), otherwise the behavior is undefined.
Note: we may not need to have a different state of lost connection (and so we could
either drop or set to a very high value m_connection_lost_count).
@timurhai: threadrunmsgcase.cpp:42, I commented a line that caused a deadlock. But is it
safe to use rendersIt.getRender() then? How else could I do? Also, I need you to check
how hack you find the new Block::reconnect()