Resolve bug #265 and introduce pixelcrop
Since H15.5, SideFX changed image:crop default value.
From SESI:
...
image:crop default value was changed when we added support for overscan and generic data windows in images.
The mantra default crop window is to have no crop window (i.e. an invalid crop window). This allows mantra to know whether a bounding box was actually specified in the IFD or not.As a note, the code in mantra looks something like:

crop = getProperty("crop");
pcrop = getProperty("pixelcrop");
screen.makeInvalid();
if (pcrop.isValid())
     screen = pcrop;
if (crop.isValid())
{
     crop.convertToPixels();
     if (screen.isValid())
         screen.intersect(crop);
     else
         screen = crop;
}
if (!screen.isValid())
     screen = [0, xres-1, 0, yres-1]
...
This commit resolve the bug #265 and introduce a crop with pixelcrop, a crop with pixel values.