afserver: solving function rewritten
For now it has the same functionality as previous.
But now its more clear and ready to be extended for new solving algorithms.
It is a step toward render pools and job projects.
Also new config parameter appeared:

*af_solving_wake_per_cycle*

It limits renders wake on lan speed. Default value is 1.
And parameter *af_serve_tasks_speed* renamed in *af_solving_tasks_speed*
it also moved in "Solving" section.