Server clients processing workflow changed.
Now server does not raise a thread on each connection.
It raises a thread pool to read/write sockets and another pool to process message.
Now there is no need to reduce threads size, but it is still configurable.

Also sockets IO is separated from other code and incapsulated in a special classes.
In feature it will be more simple to write/use some other IO workflow, non-blocking, for example epoll.
And also now it is more simple to make server to answer from run thread, to answer on actions (not only on get requests).
And this is my next goal.