afserver can return message from RUN thread.
For now any message from run thread is the same and formal.
Later server will reply on actions, to confirm changes or return an error,
for example that an action object is invalid, node to operate was not found and so on.

Note, now AfWatch works with delays, as libafqt network workflow is blocking,
but waiting afserver action response needs some time.

The next goal is to switch libafqt to non-blocking i/o.