added .clang-format file for js and cpp / reformatted common.js as example
clang-format can be used to refactor cpp and js files. The clang-format file itself should be adopted if edge cases occur which should be handled in a different way.  In a first run the whole code base should be refactored file by file and the output checked / adopted. After that clang-format can be used to easily reformat new patches to the project style.

The following cli command was used for refactoring:
```clang-format-3.8 -i -style=file common.js```