afserver: user running counts calculation fix
Now running counts calculations are removed from v_refresh() function.
I just forgot to remove it from UserAf class.
So later there was an error on a running job deletion.
UserAf::v_refresh() function ignored running counts on a job that is not running (is deleting - OFF).
When tasks are stopped, there was a double attempt to decrement counters again.