afserver: adding block/task to an existing job: action message answer contains new block/task ids.
Now if there is Action::answer_kind is not set
(an empty string, used to be 'error','log','info'),
Action::answer data will be written as a raw string in 'object' field.
This was server can return a custom JSON object containing some data as an answer on action.

Needed on appending new tasks on a dynamic Houdini PDG job.

References #451, #436.