rules: Rules server will be rewritten on Python.
It will automatically get a Python API later.

Rules server was started as a single (and simple) PHP script.
It was just historically as it was long ago.
Python web server scriptig was just started that time.
And PHP already was a mature solution for web (much before Python).

Also Python became a CG "native" language (long ago too).
Such things as project tracker should be written on Python and definitely not PHP now.